# swagger.yaml
swagger: '2.0'
info:
  version: '0.1.0'
  title: Elephancy backend (Simple)
  description: Elephancy backend (Simple)
host: example.com
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /pages:
    get:
      description: Returns all pages 
      operationId: listPages
      produces:
        - application/json
      responses:
        '200':
          description: pages response
          schema:
            type: array
            items:
              $ref: '#/definitions/page'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /pages/FindPageByPrettyURL:
    get:
      description: Find page by pretty url 
      operationId: findPageByPrettyURL
      produces:
        - application/json
      responses:
        '200':
          description: return page
          schema:
            $ref: '#/definitions/page'
        '404':
          description: page not found
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /pages/{id}:
    get:
      description: Returns a page based on a single ID
      operationId: findPageById
      produces:
        - application/json
        - application/xml
        - text/xml
        - text/html
      parameters:
        - name: id
          in: path
          description: ID of page to fetch
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: page response
          schema:
            $ref: '#/definitions/page'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
definitions:
  page:
    type: object
    required:
      - id
      - links
      - prettyurl
      - linkname
      - linkweight
      - metatitle
    properties:
      id:
        type: integer
        format: int64
      links:
        "$ref": "#/definitions/links"
      prettyurl:
        type: string
      linkname:
        type: string
      linkweight:
        type: string
      metatitle:
        type: string
  links:
    type: object
    required:
      - self
    properties:
      self:
        type: string 
  errorModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string